Resolve endless loop issue for tests on MSVCC.
It seems like there is a bug regarding spdlog and catch2. The massive calling of wmtk::logger()::warn() in `_debug_id()` caused an endless loop when running the tests compiled with MSVCC. Increasing the logger level in the `tupel_equal` function resolved that issue. It is a bit of a dirty hack but it does the job.